+++ FacturasSimplified/Invoices/DTOs/InvoiceCalculateAmountsDto.cs
@@ -2,9 +2,9 @@ namespace Facturas_simplified.Invoices.DTOs
 {
   public record InvoiceCalculatedAmountsDto
   {
-    public double Subtotal { get; init; }
-    public double TaxAmount { get; init; }
-    public double Total { get; init; }
+    public decimal Subtotal { get; init; }
+    public decimal TaxAmount { get; init; }
+    public decimal Total { get; init; }
   }
 
 }
