+++ FacturasSimplified/Invoices/DTOs/InvoiceDetailsDto.cs
@@ -9,8 +9,8 @@ namespace Facturas_simplified.Invoices.DTOs
     public int Quantity { get; set; }
 
     public required string Description { get; set; }
-    public double UnitPrice { get; set; }
-    public double SubTotal { get; set; }
+    public decimal UnitPrice { get; set; }
+    public decimal SubTotal { get; set; }
     public string? SectionTitle { get; set; }
   }
 }
